changed from %llu to %.0f in order to remain portable while retaining the ability to handle very large numbers. (double is good for about 51 bits with 1 byte accuracy. That is plenty).

+++ source/client/client.c
@@ -430,7 +430,7 @@ static void cmd_dir(void)
 
 	do_dskattr();
 
-	DEBUG(3, ("Total bytes listed: %llu\n", dir_total));
+	DEBUG(3, ("Total bytes listed: %.0f\n", (double)dir_total));
 }
